+++ src/content/components/follow.js
@@ -1,4 +1,4 @@
-import * as followActions from 'actions/follow';
+import * as followActions from 'content/actions/follow';
 import messages from 'shared/messages';
 import Hint from 'content/hint';
 import HintKeyProducer from 'content/hint-key-producer';
