+++ src/content/index.js
@@ -2,10 +2,10 @@ import './console-frame.scss';
 import * as consoleFrames from './console-frames';
 import * as settingActions from 'settings/actions/setting';
 import { createStore } from 'store';
-import ContentInputComponent from 'components/content-input';
-import KeymapperComponent from 'components/keymapper';
-import FollowComponent from 'components/follow';
-import reducers from 'reducers';
+import ContentInputComponent from 'content/components/content-input';
+import KeymapperComponent from 'content/components/keymapper';
+import FollowComponent from 'content/components/follow';
+import reducers from 'content/reducers';
 import messages from 'shared/messages';
 
 const store = createStore(reducers);
